Move internally tagged enum tests into a dedicated file
Ctrl+X, Ctrl+V
